fix(network): overriding _readableState causes weird bugs of process exiting in deadlock
When I integrate with NestJS, I could create GET requests but POST requests returned null in
response because the NodeJS exits when it could not wait an promise. This bug was caused by
request.ts when I override the internal variable of stream _readableState, with this, the NestJS
could not process the request and the process exits after some seconds without exit code.